+++ packages/plugin-commons/web/src/index.ts
@@ -31,9 +31,6 @@ export { default as createBaseInsertPluginButton } from './Base/createBaseInsert
 export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
-export { default as Image } from './Components/Image';
-export { default as SelectionList } from './Components/SelectionList';
-export { Tab, Tabs } from './Components/Tabs';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';