+++ packages/plugin-commons/web/src/index.ts
@@ -28,8 +28,6 @@ export { default as createBaseInsertPluginButton } from './Base/createBaseInsert
 export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
-export { default as Image } from './Components/Image';
-export { Tab, Tabs } from './Components/Tabs';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';