+++ packages/editor-common/web/src/Utils/draftUtils.ts
@@ -140,27 +140,30 @@ export const insertLinkAtCurrentSelection = (
 ) => {
   let selection = getSelection(editorState);
   let newEditorState = editorState;
-  if (selection.isCollapsed()) {
-    const { url } = entityData;
-    const urlToInsertWhenCollapsed = text ? text : url;
-    const contentState = Modifier.insertText(
-      editorState.getCurrentContent(),
-      selection,
-      urlToInsertWhenCollapsed
-    );
-    selection = selection.merge({
-      focusOffset: selection.getFocusOffset() + urlToInsertWhenCollapsed.length,
-    }) as SelectionState;
-    newEditorState = EditorState.push(editorState, contentState, 'insert-characters');
-  }
-  const isExistsLink = isSelectionBelongsToExistingLink(newEditorState, selection);
+  let editorStateWithLink, editorStateSelection;
   const linkEntityData = createLinkEntityData(entityData);
-  const editorStateWithLink = isExistsLink
-    ? updateLink(newEditorState, selection, linkEntityData)
-    : insertLink(newEditorState, selection, linkEntityData);
-  const editorStateSelection = isExistsLink
-    ? selection.merge({ anchorOffset: selection.getFocusOffset() })
-    : editorStateWithLink.getCurrentContent().getSelectionAfter();
+  const isExistsLink = isSelectionBelongsToExistingLink(newEditorState, selection);
+
+  if (isExistsLink) {
+    editorStateWithLink = updateLink(newEditorState, selection, linkEntityData);
+    editorStateSelection = selection.merge({ anchorOffset: selection.getFocusOffset() });
+  } else {
+    if (selection.isCollapsed()) {
+      const { url } = entityData;
+      const urlToInsertWhenCollapsed = text ? text : url;
+      const contentState = Modifier.insertText(
+        editorState.getCurrentContent(),
+        selection,
+        urlToInsertWhenCollapsed
+      );
+      selection = selection.merge({
+        focusOffset: selection.getFocusOffset() + urlToInsertWhenCollapsed.length,
+      }) as SelectionState;
+      newEditorState = EditorState.push(editorState, contentState, 'insert-characters');
+    }
+    editorStateWithLink = insertLink(newEditorState, selection, linkEntityData);
+    editorStateSelection = editorStateWithLink.getCurrentContent().getSelectionAfter();
+  }
   return EditorState.forceSelection(editorStateWithLink, editorStateSelection as SelectionState);
 };