+++ CHANGELOG.md
@@ -34,6 +34,7 @@
   - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
   - [#2604](https://github.com/wix-incubator/rich-content/pull/2604) moved TextInput component into ui-components folder
   - [#2571](https://github.com/wix-incubator/rich-content/pull/2571) moved SettingsSection component into ui-components folder
+  - [#2604](https://github.com/wix-incubator/rich-content/pull/2604) moved SelectionList/Tabs/Tab/Image components into ui-components folder
 
 ### :rocket: New Feature