+++ CHANGELOG.md
@@ -20,6 +20,10 @@
 ### :nail_care: Polish
 - `schema`
   - [#2606](https://github.com/wix-incubator/rich-content/pull/2606) code block type refactor
+### :house: Internal
+
+- `ui-components`
+  - [#2607](https://github.com/wix-incubator/rich-content/pull/2607) moved SelectionList/Tabs/Tab/Image/SliderPanel components into ui-components folder
 
 </details>
 <hr/>