+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - `editor`
   - [#2593](https://github.com/wix/ricos/pull/2593) fix cut behavior
+  - [#2603](https://github.com/wix/ricos/pull/2603) fix insertDecoration (link) from Editor Commands
 ### :nail_care: Polish
 
 - `schema`