+++ CHANGELOG.md
@@ -17,6 +17,19 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `editor`
+  - [#2593](https://github.com/wix/ricos/pull/2593) fix cut behavior
+  - [#2603](https://github.com/wix/ricos/pull/2603) fix insertDecoration (link) from Editor Commands
+### :nail_care: Polish
+
+- `schema`
+  - [#2606](https://github.com/wix-incubator/rich-content/pull/2606) code block type refactor
+### :house: Internal
+
+- `ui-components`
+  - [#2607](https://github.com/wix-incubator/rich-content/pull/2607) moved SelectionList/Tabs/Tab/Image/SliderPanel components into ui-components folder
 
 </details>
 <hr/>