+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 - `editor`
   - [#2593](https://github.com/wix/ricos/pull/2593) fix cut behavior
   - [#2603](https://github.com/wix/ricos/pull/2603) fix insertDecoration (link) from Editor Commands
+
+- `link`
+  - [#2586](https://github.com/wix/ricos/pull/2586) support anchor from outer editor into collapsible list
 ### :nail_care: Polish
 
 - `schema`